+++ src/theme/material/templates/index.ts
@@ -8,6 +8,8 @@ const templates = [
 	require('./filter-cell-currency.tpl.html'),
 	require('./filter-cell-bool.tpl.html'),
 
+	require('./column-filter-cell-text.tpl.html'),
+
 	require('./body-cell-array.tpl.html'),
 	require('./body-cell-bool.tpl.html'),
 	require('./body-cell-email.tpl.html'),