+++ CHANGELOG.md
@@ -15,6 +15,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Query builder input performance improvements.
 * Batch edit improvements.
 
+## [5.3.7] - 2018-07-08
+### Added
+- Template support of column filter items.
+
+### Fixed
+- Column sort plugin uses `FocusAfterRender` only on click.
+
 ## [5.3.6] - 2018-04-20
 ### Added
 - `Persistence` plugin - added groups.
@@ -189,8 +196,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Pager plugin.
 * Progress plugin.
 
-[unreleased]: https://github.com/qgrid/ng2/compare/v5.3.5...HEAD
-[5.3.4]: https://github.com/qgrid/ng2/compare/v5.3.5...v5.3.4
+[unreleased]: https://github.com/qgrid/ng2/compare/v5.3.7...HEAD
+[5.3.7]: https://github.com/qgrid/ng2/compare/v5.3.7...v5.3.6
+[5.3.6]: https://github.com/qgrid/ng2/compare/v5.3.6...v5.3.5
+[5.3.5]: https://github.com/qgrid/ng2/compare/v5.3.5...v5.3.4
 [5.3.4]: https://github.com/qgrid/ng2/compare/v5.3.4...v5.3.2
 [5.3.3]: https://github.com/qgrid/ng2/compare/v5.3.2...v5.2.4
 [5.3.2]: https://github.com/qgrid/ng2/compare/v5.3.2...v5.2.4